Refactor large 'export' module into parts.
